# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /arena/challenges/{challengeId}/recording-upload:
    post:
      operationId: requestArenaRecordingUpload
      tags: [arena]
      summary: Reserve a private upload for a completed match recording
      parameters:
        - in: path
          name: challengeId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArenaRecordingRequest"
      responses:
        "200":
          description: Temporary direct-upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArenaRecordingUpload"
  /arena/challenges/{challengeId}/recording:
    post:
      operationId: finishArenaRecording
      tags: [arena]
      summary: Attach an uploaded video or snapshot replay to the participant's match
      parameters:
        - in: path
          name: challengeId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArenaRecordingInput"
      responses:
        "200":
          description: Recording available to match participants
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArenaRecordingStatus"
    get:
      operationId: getArenaRecording
      tags: [arena]
      summary: Stream a match recording to an authenticated participant
      parameters:
        - in: path
          name: challengeId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Participant-only video recording or snapshot replay (video supports byte ranges)
          content:
            video/webm: {}
            video/mp4: {}
            application/vnd.arena.replay+json: {}
components:
...
  schemas:
...
    ArenaRecordingRequest:
      type: object
      required: [contentType, sizeBytes]
      properties:
        contentType: { type: string, enum: [video/webm, video/mp4, application/vnd.arena.replay+json] }
        sizeBytes: { type: integer, minimum: 1, maximum: 40000000 }
    ArenaRecordingUpload:
      type: object
      required: [uploadUrl, objectPath]
      properties:
        uploadUrl: { type: string, format: uri }
        objectPath: { type: string }
    ArenaRecordingInput:
      type: object
      required: [objectPath, contentType]
      properties:
        objectPath: { type: string }
        contentType: { type: string, enum: [video/webm, video/mp4, application/vnd.arena.replay+json] }
    ArenaRecordingStatus:
      type: object
      required: [available, videoUrl]
      properties:
        available: { type: boolean }
        videoUrl: { type: string }
...